+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 v3.0.1
 ------
 - Added RenderContext::StateBindFlags, which allows the user to control which part of the `GraphicsState` will be bound to the pipeline
-- Added helper functions to initialize D3D12 state-objects descs (root-signature, )
+- Added helper functions to initialize D3D12 state-objects descs (root-signature, GSO-desc)